+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Added `emailVerified`, `phoneNumber` and `phoneNumberVerified` values to `UserProfile`
+
 ## [0.4.3] - 2022-08-28
 
 ## [0.4.2] - 2022-07-29